release: xugu-xorm v1.3.2-xugu
- Bump xorm dependency to v1.3.2
- Compatibility verified against xorm v1.3.1..v1.3.2
- Adapt to schemas.Column.Length/Length2 int->int64 (xorm #2160)
- All tests passed (26 unit + 18 integration + 13 realdb)

+++ xugu_test.go
@@ -177,7 +177,7 @@ func TestXuguGenScanResultUnsignedStripping(t *testing.T) {
 func TestXuguSQLType(t *testing.T) {
 	d := newXuguDialectInit(t, "test")
 
-	newCol := func(name string, l1, l2 int) *schemas.Column {
+	newCol := func(name string, l1, l2 int64) *schemas.Column {
 		return &schemas.Column{
 			Name:    name,
 			SQLType: schemas.SQLType{Name: name},
